# docs/api-snapshot/openapi.json
{
  "components": {
    "schemas": {
...
      "CreateNodeBody": {
        "properties": {
          "body": {
            "title": "Body",
            "type": "string"
          },
          "facets": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/Facet"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Facets"
          },
          "message": {
            "default": "",
            "title": "Message",
            "type": "string"
          },
          "node_type": {
            "title": "Node Type",
            "type": "string"
          },
          "task_state": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Task State"
          }
        },
        "required": [
          "node_type",
          "body"
        ],
        "title": "CreateNodeBody",
        "type": "object"
      },
...
      "DeleteNodeBody": {
        "properties": {
          "redirect_to": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Redirect To"
          },
          "tombstone": {
            "default": false,
            "title": "Tombstone",
            "type": "boolean"
          }
        },
        "title": "DeleteNodeBody",
        "type": "object"
      },
      "Facet": {
        "properties": {
          "facet_id": {
            "title": "Facet Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "span": {
            "title": "Span",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "facet_id",
          "name",
          "span",
          "version"
        ],
        "title": "Facet",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "MergeBody": {
        "properties": {
          "ids": {
            "items": {
              "type": "string"
            },
            "title": "Ids",
            "type": "array"
          }
        },
        "required": [
          "ids"
        ],
        "title": "MergeBody",
        "type": "object"
      },
      "PatchNodeBody": {
        "properties": {
          "body": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Body"
          },
          "change_class": {
            "title": "Change Class",
            "type": "string"
          },
          "facets": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/Facet"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Facets"
          },
          "facets_touched": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Facets Touched",
            "type": "array"
          },
          "message": {
            "default": "",
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "change_class"
        ],
        "title": "PatchNodeBody",
        "type": "object"
      },
...
      "SplitBody": {
        "properties": {
          "parts": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Parts",
            "type": "array"
          }
        },
        "required": [
          "parts"
        ],
        "title": "SplitBody",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "tm-daemon API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/nodes": {
      "post": {
        "operationId": "create_node_v1_nodes_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateNodeBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Create Node V1 Nodes Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Node",
        "tags": [
          "nodes"
        ]
      }
    },
    "/v1/nodes/{node_id}": {
      "delete": {
        "operationId": "delete_node_v1_nodes__node_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "node_id",
            "required": true,
            "schema": {
              "title": "Node Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/DeleteNodeBody"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Payload"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Delete Node V1 Nodes  Node Id  Delete",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Node",
        "tags": [
          "nodes"
        ]
      },
      "get": {
        "operationId": "get_node_v1_nodes__node_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "node_id",
            "required": true,
            "schema": {
              "title": "Node Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "as_of",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "As Of"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Get Node V1 Nodes  Node Id  Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Node",
        "tags": [
          "nodes"
        ]
      },
      "patch": {
        "operationId": "patch_node_v1_nodes__node_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "node_id",
            "required": true,
            "schema": {
              "title": "Node Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatchNodeBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Patch Node V1 Nodes  Node Id  Patch",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Patch Node",
        "tags": [
          "nodes"
        ]
      }
    },
    "/v1/nodes/{node_id}/history": {
      "get": {
        "operationId": "node_history_v1_nodes__node_id__history_get",
        "parameters": [
          {
            "in": "path",
            "name": "node_id",
            "required": true,
            "schema": {
              "title": "Node Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Node History V1 Nodes  Node Id  History Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Node History",
        "tags": [
          "nodes"
        ]
      }
    },
    "/v1/nodes/{node_id}/merge": {
      "post": {
        "operationId": "merge_nodes_v1_nodes__node_id__merge_post",
        "parameters": [
          {
            "in": "path",
            "name": "node_id",
            "required": true,
            "schema": {
              "title": "Node Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MergeBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Merge Nodes V1 Nodes  Node Id  Merge Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Merge Nodes",
        "tags": [
          "nodes"
        ]
      }
    },
    "/v1/nodes/{node_id}/neighborhood": {
      "get": {
        "operationId": "node_neighborhood_v1_nodes__node_id__neighborhood_get",
        "parameters": [
          {
            "in": "path",
            "name": "node_id",
            "required": true,
            "schema": {
              "title": "Node Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "hops",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 0,
              "title": "Hops",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Node Neighborhood V1 Nodes  Node Id  Neighborhood Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Node Neighborhood",
        "tags": [
          "nodes"
        ]
      }
    },
    "/v1/nodes/{node_id}/split": {
      "post": {
        "operationId": "split_node_v1_nodes__node_id__split_post",
        "parameters": [
          {
            "in": "path",
            "name": "node_id",
            "required": true,
            "schema": {
              "title": "Node Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SplitBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Split Node V1 Nodes  Node Id  Split Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Split Node",
        "tags": [
          "nodes"
        ]
      }
    },
    "/v1/nodes/{node_id}/vet": {
      "post": {
        "operationId": "vet_node_v1_nodes__node_id__vet_post",
        "parameters": [
          {
            "in": "path",
            "name": "node_id",
            "required": true,
            "schema": {
              "title": "Node Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Vet Node V1 Nodes  Node Id  Vet Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Vet Node",
        "tags": [
          "nodes"
        ]
      }
    },
...
  }
}